test of sane project setup passed
