+++ api/tests/__init__.py
@@ -1,1 +1,4 @@
-from .ticket_test import TicketTestCase, TicketViewTestCase 
+from .member_test import MemberBaseBehavior
+from .profile_test import ProfileBaseBehavior
+from .team_test import TeamBaseBehavior
+from .ticket_test import TicketTestCase, TicketViewTestCase
